requirements.txt
asgiref==3.5.2
crispy-bootstrap5==0.7
Django==4.1.4
django-crispy-forms==1.14.0
django-extra-views==0.14.0
Faker==15.3.4
python-dateutil==2.8.2
six==1.16.0
sqlparse==0.4.3
sweetify==2.3.1
tzdata==2022.7
python-dotenv~=1.0.0